i182: move serverless example to hands-on
