+++ hands-on/serverless/.terraform.lock.hcl
