+++ hands-on/serverless/README.md
@@ -1,0 +1,6 @@
+# Architecture:
+[User] -> [API_Gateway] -> [Lambda] -> [DynamoDB]
+
+# References:
+## https://developer.hashicorp.com/terraform/tutorials/aws/lambda-api-gateway
+## https://dev.to/rolfstreefkerk/openapi-with-terraform-on-aws-api-gateway-17je
