+++ hands-on/serverless/test_data.json
@@ -1,0 +1,7 @@
+{
+  "operation": "echo",
+  "payload": {
+      "somekey1": "somevalue1",
+      "somekey2": "somevalue2"
+  }
+}
