--- serverless/test_data.json
@@ -1,7 +1,0 @@
-{
-  "operation": "echo",
-  "payload": {
-      "somekey1": "somevalue1",
-      "somekey2": "somevalue2"
-  }
-}
